% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_IBNR.R
\name{calculate_IBNR}
\alias{calculate_IBNR}
\title{calculate IBNR claims}
\usage{
calculate_IBNR(.data, .ibnrtable, .ibnrfactor, .cy)
}
\arguments{
\item{.data}{A data frame coming out from the function \code{optimize_exposure()}, or of the same type.}

\item{.ibnrtable}{A table with two columns, of the for IBNR factors and one for the associated calendar
year.}

\item{.ibnrfactor}{the column of .ibnrtable with the IBNR factors.}

\item{.cy}{the column of .ibnrtable with the corresponding calendar year of the IBNR factors.}
}
\value{
A new data frame with new columns:
\itemize{
\item \code{IBNR_factor}, a column that indicates the IBNR factor of the corresponding current year;
\item \code{CLAIMS_IBNR}, a column with the associated IBNR claim for each \code{CY} and policyholder (NOTE: if the
current year in the original data frame is higher than the most recent current year in the IBNR table,
this will produce an \code{NA}.)
}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\description{
given as input a data frame of insurance policies coming out from the function \code{optimize_exposure()},
and a table with IBNR factors, calculate the IBNR claims.
}
\examples{
data(lifepolicies)

lifepolicies \%>\%
  optimize_exposure(
    .polnr = Policy_id, .inceptiondate = Inception_date,
    .termdate = Term_date, .claimdate = Claim_date,
    .lapsedate = Lapse_date, .birthdate = Birth_date,
    .gender = Gender
  ) \%>\%
  compute_expectedclaim(
    .table = SI2006, .age = AGE,
    .gender = GENDER, .qx = qx
  ) \%>\%
  calculate_IBNR(
    .ibnrtable = IBNR_factors_CQP,
    .ibnrfactor = IBNR_factor, .cy = CY
  )
}
