% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_date.R
\name{check_date}
\alias{check_date}
\title{Check if column class is date}
\usage{
check_date(.df, .var)
}
\arguments{
\item{.df}{a dataframe.}

\item{.var}{the name of the column to test.}
}
\value{
the function returns TRUE if .var is of class "Date", FALSE otherwise.
}
\description{
function to check whether a given column of the dataset is of the class "Date".
}
\examples{
check_date(.df = lifepolicies, .var = Term_date)
}
