% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_event_cols.R
\name{check_events_col}
\alias{check_events_col}
\title{check the events columns}
\usage{
check_events_col(.df, .var)
}
\arguments{
\item{.df}{a dataframe.}

\item{.var}{the name of the column to test.}
}
\value{
the function returns an error message if the condition is not satisfied.
}
\description{
check that the columns representing the event (claim or lapse) are made up of ones and zeros,
and that are integer values.
}
