% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_expectedclaim.R
\name{compute_expectedclaim}
\alias{compute_expectedclaim}
\title{Compute the expected claims}
\usage{
compute_expectedclaim(.data, .table, .age, .gender, .qx)
}
\arguments{
\item{.data}{data frame in the format provided by \code{optimize_exposure()}.}

\item{.table}{a table with the probabilities qx.}

\item{.age}{the .table's column with the age associated to qx.}

\item{.gender}{the .table's column with the gender associated with each qx.}

\item{.qx}{the .table's column with the probabilities qx (note that it has to be per 1000).}
}
\value{
A new data frame with 4 new columns:
\itemize{
\item \code{qx.x} indicates the probability of dying between \code{AGE0} and \code{AGE1} for the policyholder;
\item \code{qx.y} indicates the probability of dying between \code{AGE1} and \code{AGE1}+ 1 for the policyolder;
\item \code{EXPECTED_CLAIM_AGE0} is the expected claim for the policyholder at \code{AGE0}, computed as
the exposure in year when he/she has \code{AGE0} * (\code{qx.x}/ 1000);
\item \code{EXPECTED_CLAIM_AGE1} is the expected claim for the policyholder at \code{AGE1}, computed as
the exposure in year when he/she has \code{AGE1} * (\code{qx.y}/ 1000).
}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\description{
Given as input a data frame of insurance policies coming out from \code{optimize_exposure()}, it
returns a new data frame with the associated expected claim for each age of the policyhoder.
}
\examples{
data(lifepolicies)

lifepolicies \%>\%
  optimize_exposure(
  .polnr = Policy_id, .inceptiondate = Inception_date,
  .termdate = Term_date, .mainevent = Claim , .mainevent_date = Claim_date,
  .subevent = Lapse, .subevent_date = Lapse_date, .birthdate = Birth_date,
  .gender = Gender, .waitingperiod = Waiting_period
  ) \%>\%
  compute_expectedclaim(
    .table = SI2006, .age = AGE,
    .gender = GENDER, .qx = qx
  )
}
