% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_portfolio.R
\name{duplicate_portfolio}
\alias{duplicate_portfolio}
\title{Duplicate the rows of an insurance portfolio}
\usage{
duplicate_portfolio(.data, .polnr)
}
\arguments{
\item{.data}{A data frame of insurance policies with n rows for each policy, where n is the number
of years the contract is in force.}

\item{.polnr}{A column of .data with the policy identification numbers.}
}
\value{
A new data frame with n times 2 rows for each policy. This type of dataset is optimal to
use the function \code{optimize_exposure()}.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\description{
Given a dataset with n rows for each policy, where n is the number of years the policy is in force,
the function duplicate the dataset, providing a new object that is suited to the function \code{optimize_exposure()}.
}
\examples{

data(raw_lifepolicies)

duplicate_lifepolicies <- duplicate_portfolio(.data = raw_lifepolicies, .polnr = Policy_id)
}
