% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_portfolio.R
\name{expand_portfolio}
\alias{expand_portfolio}
\title{Expand an insurance portfolio.}
\usage{
expand_portfolio(
  .data,
  .polnr,
  .claim,
  .claimdate,
  .inceptiondate,
  .termdate,
  .lapse,
  .lapsedate
)
}
\arguments{
\item{.data}{A data frame of insurance policies that has one single row for each policy.}

\item{.polnr}{A column of .data with the policy identification numbers.}

\item{.claim}{A column of .data with 1 if the claim occurred, 0 otherwise.}

\item{.claimdate}{A column of .data with the date of the claim, if any, otherwise 2099-12-31.}

\item{.inceptiondate}{A column of .data with the starting date of the policies.}

\item{.termdate}{A column of .data with the expiry date of the policies.}

\item{.lapse}{A column of .data with 1 if the policy is lapsed, 0 otherwise.}

\item{.lapsedate}{A column of .data with the date of the lapse, if any, otherwise 2099-12-31.}
}
\value{
A dataset with n x 2 rows for each policy, where n is the number of years the contract is in force.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\description{
Given a dataset of insurance policies with one row for each of them, the function first
expands it in such a way that each policy has n rows, where n is the number of years the contract is
in force, then it duplicates the rows, so that each policy has n x 2 rows. In this way the dataset is
suited to be passed to the function \code{optimize_exposure()}.
}
\examples{
data(onerow_lifepolicies)

onerow_lifepolicies \%>\%
  expand_portfolio(
    .polnr = Policy_id, .claim = Claim,
    .claimdate = Claim_date, .inceptiondate = Inception_date,
    .termdate = Term_date, .lapsedate = Lapse_date, .lapse = Lapse
  )
}
