% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IBNR_factors_CQP}
\alias{IBNR_factors_CQP}
\title{IBNR Factors}
\format{
A data frame with variables:
\describe{
\item{CY}{A column with the current year associated with the IBNR factor.}
\item{IBNR_factor}{A column with the IBNR factors.}
}
}
\usage{
IBNR_factors_CQP
}
\description{
A table with the IBNR (Incurred but not reported) claim factors for each current year.
}
\examples{
data(IBNR_factors_CQP)
head(IBNR_factors_CQP)
}
\keyword{datasets}
