% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{onerow_lifepolicies}
\alias{onerow_lifepolicies}
\title{Life policies portfolio (one row for each policy).}
\format{
A data frame with variables:
\describe{
\item{Policy_id}{policy number identifier, unique for each policy.}
\item{Inception_date}{signing date of the contract.}
\item{Term_date}{ending date of the contract.}
\item{Birth_date}{date of birth of the policyolder.}
\item{Gender}{policyholder's gender.}
\item{UY}{underwriting year.}
\item{Claim_date}{date when claims occurred, if any, otherwise 2099-12-31.}
\item{Claim}{1 if the year of \code{Claim_date} is equal to CY, 0 otherwise.}
\item{Lapse_date}{date when policy lapsed, if so, otherwise 2099-12-31.}
\item{Lapse}{1 if the year of \code{Lapse_date} is equal to CY, 0 otherwise.}
}
}
\usage{
onerow_lifepolicies
}
\description{
Simulated life insurance policies portfolio made up of 2,500 policies.
Note that the dataset is not suited to be passed in the function \code{optimize_exposure()}. This because
it has one row for each policy.
This dataset needs to be expanded using the function \code{expand_portfolio()} before computing the
optimal exposure with \code{optimize_exposure()}.
}
\examples{
data(onerow_lifepolicies)
head(onerow_lifepolicies)
}
\keyword{datasets}
