% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_exposure.R
\name{optimize_exposure}
\alias{optimize_exposure}
\title{Compute optimal exposure}
\usage{
optimize_exposure(
  .data,
  .polnr,
  .inceptiondate,
  .termdate,
  .claimdate,
  .lapsedate,
  .birthdate,
  .gender
)
}
\arguments{
\item{.data}{A data frame of insurance policies that has a number of rows for each policy equal to:
year in force of the policy x 2. This type of format is required to compute in an optimal
way the exposure in each calendar year for each age of the policyholder.}

\item{.polnr}{A column of .data with the policy identification numbers.}

\item{.inceptiondate}{A column of .data with the starting date of the policies.}

\item{.termdate}{A column of .data with the expiry date of the policies.}

\item{.claimdate}{A column of .data with the date of the claim, if any, otherwise 2099-12-31.}

\item{.lapsedate}{A column of .data with the date of the lapse, if any, otherwise 2099-12-31.}

\item{.birthdate}{A column of .data with the birthday of the policyholders.}

\item{.gender}{A column of .data with the gender of the policyholders.}
}
\value{
The modified data frame with some new columns, in particular:
\itemize{
\item \code{inception_valuation_date} indicates the date from which to start to compute the exposure,
it can be the inception date, the renewal date or the first of January of the current year;
\item \code{end_valuation_date} indicates the ending date to take into account in the computation of
the exposure, it can be the last day of the current year, the claim/lapse date or the renewal
date (this because in the following row the inception_valuation_date will be the same);
\item \code{AGE0} is the age of the policyholder at the inception valuation date;
\item \code{AGE1} is the possible age of the policyholder at the end valuation date;
\item \code{DAY_AGE0} indicates the day of exposure of the policyholder with \code{AGE0}(before his/her anniversary)
in the valuation period;
\item \code{DAY_AGE1} indicates the day of exposure of the policyholder with \code{AGE1} (after his/her anniversary)
in the valuation period;
\item \code{EXPOSURE_YEAR_AGE0} is the exposure of the policyholder in the current year when he/she has \code{AGE0},
(it is computed as number days exposed with age0 / total number of days in the current year);
\item \code{EXPOSURE_YEAR_AGE1} is the exposure of the policyholder in the current year when he/she has \code{AGE1},
(it is computed as number days exposed with age1 / total number of days in the current year);
\item \code{EXPOSURE_IN_YEAR} is the total exposure in the valuation period, the sum of the previous two exposures,
it is 1 by definition when a claim occurs;
\item \code{EXPOSURE_IN_YEAR_lapse} is the total exposure in the valuation period, the sum of the previous two exposures,
it is 1 by definition when a policy lapses.
}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\description{
The idea behind this function is to compute in an optimal way the exposure period
for each policyhoder and for each year in which the contract is in force, by distinguishing
each calendar year in before and after his/her birthday. Doing so, for a given policyholder
and for each year we get the exposure he/she is exposed with age x and the exposire he/she
is exposed at age x+1.
}
\examples{
data(lifepolicies)
lifepolicies \%>\%
  optimize_exposure(
    .polnr = Policy_id, .inceptiondate = Inception_date,
    .termdate = Term_date, .claimdate = Claim_date,
    .lapsedate = Lapse_date, .birthdate = Birth_date,
    .gender = Gender
  )
}
