% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_lifepolicies}
\alias{raw_lifepolicies}
\title{Life policies portfolio (n rows for each policy)}
\format{
A data frame with variables:
\describe{
\item{Policy_id}{policy number identifier, unique for each policy.}
\item{Inception_date}{signing date of the contract.}
\item{Waiting_period}{The period in months, starting from signing, for which the policyholder is not covered.}
\item{Term_date}{ending date of the contract.}
\item{Birth_date}{date of birth of the policyolder.}
\item{Gender}{policyholder's gender.}
\item{UY}{underwriting year.}
\item{CY}{current year of evaluation.}
\item{Claim_date}{date when claims occurred, if any, otherwise 2099-12-31.}
\item{Claim}{1 if the year of \code{Claim_date} is equal to CY, 0 otherwise.}
\item{Lapse_date}{date when policy lapsed, if so, otherwise 2099-12-31.}
\item{Lapse}{1 if the year of \code{Lapse_date} is equal to CY, 0 otherwise.}
}#'
}
\usage{
raw_lifepolicies
}
\description{
Simulated life insurance policies portfolio made up of 2,500 policies.
Note that the dataset is not suited to be passed in the function \code{optimize_exposure()}. This because
it has n number of rows for each policy, where n is the number of years the contract is in force.
This dataset needs to be duplicated using the function \code{duplicate_portfolio()} before computing the
optimal exposure with \code{optimize_exposure()}.
}
\examples{
data(raw_lifepolicies)
head(raw_lifepolicies)
}
\keyword{datasets}
