% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SI2006}
\alias{SI2006}
\title{Sample Mortality table (2006)}
\format{
A data frame with variables:
\describe{
\item{AGE}{Age column.}
\item{GENDER}{Gender column.}
\item{qx}{ Mortality rate between age x and (x +1).
That is, the probability that a person aged x will die before reaching age (x +1).}
}
}
\usage{
SI2006
}
\description{
A Mortality table (2006) for Italian population.
}
\examples{
data(SI2006)
head(SI2006)
}
\keyword{datasets}
